Official release of 1.70.0

+++ src/main/java/net/pms/dlna/DLNAMediaInfo.java
@@ -18,15 +18,43 @@
  */
 package net.pms.dlna;
 
-import com.sun.jna.Platform;
+import java.awt.Color;
+import java.awt.Font;
+import java.awt.Graphics;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.ListIterator;
+import java.util.Locale;
+import java.util.Map;
+import java.util.NoSuchElementException;
+import java.util.StringTokenizer;
+
+import javax.imageio.ImageIO;
+
+import net.coobird.thumbnailator.Thumbnails;
+import net.coobird.thumbnailator.Thumbnails.Builder;
 import net.pms.PMS;
 import net.pms.configuration.RendererConfiguration;
 import net.pms.formats.AudioAsVideo;
 import net.pms.formats.Format;
+import net.pms.formats.v2.SubtitleType;
 import net.pms.io.OutputParams;
 import net.pms.io.ProcessWrapperImpl;
 import net.pms.network.HTTPResource;
-import net.pms.util.*;
+import net.pms.util.AVCHeader;
+import net.pms.util.CoverUtil;
+import net.pms.util.FileUtil;
+import net.pms.util.MpegUtil;
+import net.pms.util.ProcessUtil;
+
 import org.apache.commons.lang.StringUtils;
 import org.apache.sanselan.ImageInfo;
 import org.apache.sanselan.Sanselan;
@@ -42,11 +70,8 @@ import org.jaudiotagger.tag.Tag;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.imageio.ImageIO;
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.*;
-import java.util.*;
+import com.sun.jna.Platform;
+
 
 /**
  * This class keeps track of scanned MediaInfo library information.
@@ -101,6 +126,8 @@ public class DLNAMediaInfo implements Cloneable {
 	@Deprecated
 	public String frameRate;
 
+	private String frameRateMode;
+
 	/**
 	 * @deprecated Use standard getter and setter to access this variable.
 	 */
@@ -125,17 +152,8 @@ public class DLNAMediaInfo implements Cloneable {
 	@Deprecated
 	public int bitsPerPixel;
 
-	/**
-	 * @deprecated Use standard getter and setter to access this variable.
-	 */
-	@Deprecated
-	public ArrayList<DLNAMediaAudio> audioCodes = new ArrayList<DLNAMediaAudio>();
-
-	/**
-	 * @deprecated Use standard getter and setter to access this variable.
-	 */
-	@Deprecated
-	public ArrayList<DLNAMediaSubtitle> subtitlesCodes = new ArrayList<DLNAMediaSubtitle>();
+	private List<DLNAMediaAudio> audioTracks = new ArrayList<DLNAMediaAudio>();
+	private List<DLNAMediaSubtitle> subtitleTracks = new ArrayList<DLNAMediaSubtitle>();
 
 	/**
 	 * @deprecated Use standard getter and setter to access this variable.
@@ -402,53 +420,6 @@ public class DLNAMediaInfo implements Cloneable {
 		return pw;
 	}
 
-	private ProcessWrapperImpl getImageMagickThumbnail(InputFile media) throws IOException {
-	// convert -size 320x180  hatching_orig.jpg  -auto-orient -thumbnail 160x90   -unsharp 0x.5  thumbnail.gif
-		String args [] = new String[10];
-		args[0] = PMS.getConfiguration().getIMConvertPath();
-		args[1] = "-size";
-		args[2] = "320x180";
-		if (media.getFile() != null)
-			args[3] = ProcessUtil.getShortFileNameIfWideChars(media.getFile().getAbsolutePath());
-		else
-			args[3] = "-";
-		args[4] = "-auto-orient";
-		args[5] = "-thumbnail";
-		args[6] = "160x90";
-		args[7] = "-unsharp";
-		args[8] = "-0x.5";
-		args[9] = PMS.getConfiguration().getTempFolder() + "/imagemagick_thumbs/" + media.getFile().getName() + ".jpg";
-		OutputParams params = new OutputParams(PMS.getConfiguration());
-		params.workDir = new File(PMS.getConfiguration().getTempFolder().getAbsolutePath() + "/imagemagick_thumbs/");
-
-		if (!params.workDir.exists() && !params.workDir.mkdirs()) {
-			logger.debug("Could not create directory \"" + params.workDir.getAbsolutePath() + "\"");
-		}
-
-		params.maxBufferSize = 1;
-		params.stdin = media.getPush();
-		params.log = true;
-		params.noexitcheck = true; // not serious if anything happens during the thumbnailer
-		final ProcessWrapperImpl pw = new ProcessWrapperImpl(args, params);
-			// FAILSAFE
-		setParsing(true);
-		Runnable r = new Runnable() {
-			public void run() {
-				try {
-					Thread.sleep(7000);
-					ffmpeg_failure = true;
-				} catch (InterruptedException e) {}
-				pw.stopProcess();
-				setParsing(false);
-			}
-		};
-		Thread failsafe = new Thread(r, "ImageMagick Thumbnail Failsafe");
-		failsafe.start();
-		pw.runInSameThread();
-		setParsing(false);
-		return pw;
-	}
-
 	private String getFfmpegPath() {
 		String value = PMS.getConfiguration().getFfmpegPath();
 		if (value == null) {
@@ -502,13 +473,13 @@ public class DLNAMediaInfo implements Cloneable {
 							audio.setSampleFrequency("" + rate);
 							setDuration((double) length);
 							setBitrate((int) ah.getBitRateAsNumber());
-							audio.setNrAudioChannels(2);
+							audio.getAudioProperties().setNumberOfChannels(2);
 							if (ah.getChannels() != null && ah.getChannels().toLowerCase().contains("mono")) {
-								audio.setNrAudioChannels(1);
+								audio.getAudioProperties().setNumberOfChannels(1);
 							} else if (ah.getChannels() != null && ah.getChannels().toLowerCase().contains("stereo")) {
-								audio.setNrAudioChannels(2);
+								audio.getAudioProperties().setNumberOfChannels(2);
 							} else if (ah.getChannels() != null) {
-								audio.setNrAudioChannels(Integer.parseInt(ah.getChannels()));
+								audio.getAudioProperties().setNumberOfChannels(Integer.parseInt(ah.getChannels()));
 							}
 							audio.setCodecA(ah.getEncodingType().toLowerCase());
 							if (audio.getCodecA().contains("(windows media")) {
@@ -550,7 +521,7 @@ public class DLNAMediaInfo implements Cloneable {
 						audio.setSongname(f.getFile().getName());
 					}
 					if (!ffmpeg_parsing) {
-						getAudioCodes().add(audio);
+						getAudioTracksList().add(audio);
 					}
 				}
 			}
@@ -599,28 +570,49 @@ public class DLNAMediaInfo implements Cloneable {
 					// ffmpeg_parsing = true;
 					logger.info("Error parsing image with Sanselan... switching to FFmpeg: " + e.getMessage());
 				}
+
 				try {
-					if(PMS.getConfiguration().getImageThumbnailsEnabled()) {
-						getImageMagickThumbnail(f);
-						String frameName = PMS.getConfiguration().getTempFolder() + "/imagemagick_thumbs/" + f.getFile().getName() + ".jpg";
-						File jpg = new File(frameName);
-						if (jpg.exists()) {
-							InputStream is = new FileInputStream(jpg);
-							int sz = is.available();
-							if (sz > 0) {
-								setThumb(new byte [sz]);
-								is.read(getThumb());
+					if (PMS.getConfiguration().getImageThumbnailsEnabled()) {
+						File thumbDir = new File(PMS.getConfiguration().getTempFolder() + "/thumbs/");
+
+						if (!thumbDir.exists() && !thumbDir.mkdirs()) {
+							logger.debug("Could not create directory \"" + thumbDir.getAbsolutePath() + "\"");
+						} else {
+							String thumbFilename = thumbDir + f.getFile().getName() + ".jpg";
+
+							logger.trace("Creating thumbnail \"" + thumbFilename + "\"");
+
+							// Create the thumbnail image using the Thumbnailator library
+							final Builder<File> thumbnail = Thumbnails.of(f.getFile());
+							thumbnail.size(320, 180);
+							thumbnail.outputFormat("jpg");
+							thumbnail.outputQuality(1.0f);
+							thumbnail.toFile(thumbFilename);
+
+							File jpg = new File(thumbFilename);
+
+							if (jpg.exists()) {
+								InputStream is = new FileInputStream(jpg);
+								int sz = is.available();
+
+								if (sz > 0) {
+									setThumb(new byte[sz]);
+									is.read(getThumb());
+								}
+
+								is.close();
+
+								if (!jpg.delete()) {
+									jpg.deleteOnExit();
+								}
 							}
-							is.close();
-							if (!jpg.delete())
-								jpg.deleteOnExit();
 						}
 					}
-				} catch (Throwable e) {
-					logger.info("Error generating thumbnail of image with ImageMagick: " + e.getMessage());
-				
+				} catch (Exception e) {
+					logger.info("Error generating thumbnail of image", e);
 				}
 			}
+
 			if (ffmpeg_parsing) {
 				if (!thumbOnly || !PMS.getConfiguration().isUseMplayerForVideoThumbs()) {
 					pw = getFFMpegThumbnail(f);
@@ -691,9 +683,6 @@ public class DLNAMediaInfo implements Cloneable {
 								int a = line.indexOf("(");
 								int b = line.indexOf("):", a);
 								DLNAMediaAudio audio = new DLNAMediaAudio();
-								if (langId == 0 && (getContainer().equals("avi") || getContainer().equals("ogm") || getContainer().equals("mov") || getContainer().equals("flv") || getContainer().equals("mp4"))) {
-									langId++;
-								}
 								audio.setId(langId++);
 								if (a > -1 && b > a) {
 									audio.setLang(line.substring(a + 1, b));
@@ -720,17 +709,17 @@ public class DLNAMediaInfo implements Cloneable {
 									} else if (token.endsWith("Hz")) {
 										audio.setSampleFrequency(token.substring(0, token.indexOf("Hz")).trim());
 									} else if (token.equals("mono")) {
-										audio.setNrAudioChannels(1);
+										audio.getAudioProperties().setNumberOfChannels(1);
 									} else if (token.equals("stereo")) {
-										audio.setNrAudioChannels(2);
+										audio.getAudioProperties().setNumberOfChannels(2);
 									} else if (token.equals("5:1") || token.equals("5.1") || token.equals("6 channels")) {
-										audio.setNrAudioChannels(6);
+										audio.getAudioProperties().setNumberOfChannels(6);
 									} else if (token.equals("5 channels")) {
-										audio.setNrAudioChannels(5);
+										audio.getAudioProperties().setNumberOfChannels(5);
 									} else if (token.equals("4 channels")) {
-										audio.setNrAudioChannels(4);
+										audio.getAudioProperties().setNumberOfChannels(4);
 									} else if (token.equals("2 channels")) {
-										audio.setNrAudioChannels(2);
+										audio.getAudioProperties().setNumberOfChannels(2);
 									} else if (token.equals("s32")) {
 										audio.setBitsperSample(32);
 									} else if (token.equals("s24")) {
@@ -758,7 +747,7 @@ public class DLNAMediaInfo implements Cloneable {
 										}
 									}
 								}
-								getAudioCodes().add(audio);
+								getAudioTracksList().add(audio);
 							} else if (line.indexOf("Video:") > -1) {
 								StringTokenizer st = new StringTokenizer(line, ",");
 								while (st.hasMoreTokens()) {
@@ -802,7 +791,7 @@ public class DLNAMediaInfo implements Cloneable {
 								}
 							} else if (line.indexOf("Subtitle:") > -1 && !line.contains("tx3g")) {
 								DLNAMediaSubtitle lang = new DLNAMediaSubtitle();
-								lang.setType((line.contains("dvdsub") && Platform.isWindows() ? DLNAMediaSubtitle.VOBSUB : DLNAMediaSubtitle.EMBEDDED));
+								lang.setType((line.contains("dvdsub") && Platform.isWindows() ? SubtitleType.VOBSUB : SubtitleType.UNKNOWN));
 								int a = line.indexOf("(");
 								int b = line.indexOf("):", a);
 								if (a > -1 && b > a) {
@@ -830,7 +819,7 @@ public class DLNAMediaInfo implements Cloneable {
 										}
 									}
 								}
-								getSubtitlesCodes().add(lang);
+								getSubtitleTracksList().add(lang);
 							}
 						}
 					}
@@ -1073,14 +1062,14 @@ public class DLNAMediaInfo implements Cloneable {
 
 	public String toString() {
 		String s = "container: " + getContainer() + " / bitrate: " + getBitrate() + " / size: " + getSize() + " / codecV: " + getCodecV() + " / duration: " + getDurationString() + " / width: " + getWidth() + " / height: " + getHeight() + " / frameRate: " + getFrameRate() + " / thumb size : " + (getThumb() != null ? getThumb().length : 0) + " / muxingMode: " + getMuxingMode();
-		for (DLNAMediaAudio audio : getAudioCodes()) {
-			s += "\n\taudio: id=" + audio.getId() + " / lang: " + audio.getLang() + " / flavor: " + audio.getFlavor() + " / codec: " + audio.getCodecA() + " / sf:" + audio.getSampleFrequency() + " / na: " + audio.getNrAudioChannels() + " / bs: " + audio.getBitsperSample();
+		for (DLNAMediaAudio audio : getAudioTracksList()) {
+			s += "\n\taudio: id=" + audio.getId() + " / lang: " + audio.getLang() + " / flavor: " + audio.getFlavor() + " / codec: " + audio.getCodecA() + " / sf:" + audio.getSampleFrequency() + " / na: " + (audio.getAudioProperties() != null ? audio.getAudioProperties().getNumberOfChannels() : "-") + " / bs: " + audio.getBitsperSample();
 			if (audio.getArtist() != null) {
 				s += " / " + audio.getArtist() + "|" + audio.getAlbum() + "|" + audio.getSongname() + "|" + audio.getYear() + "|" + audio.getTrack();
 			}
 		}
-		for (DLNAMediaSubtitle sub : getSubtitlesCodes()) {
-			s += "\n\tsub: id=" + sub.getId() + " / lang: " + sub.getLang() + " / flavor: " + sub.getFlavor() + " / type: " + sub.getType();
+		for (DLNAMediaSubtitle sub : getSubtitleTracksList()) {
+			s += "\n\tsub: id=" + sub.getId() + " / lang: " + sub.getLang() + " / flavor: " + sub.getFlavor() + " / type: " + (sub.getType() != null ? sub.getType().toString() : "null");
 		}
 		return s;
 	}
@@ -1124,8 +1113,8 @@ public class DLNAMediaInfo implements Cloneable {
 	}
 
 	public DLNAMediaAudio getFirstAudioTrack() {
-		if (getAudioCodes().size() > 0) {
-			return getAudioCodes().get(0);
+		if (getAudioTracksList().size() > 0) {
+			return getAudioTracksList().get(0);
 		}
 		return null;
 	}
@@ -1271,13 +1260,13 @@ public class DLNAMediaInfo implements Cloneable {
 		Object cloned = super.clone();
 		if (cloned instanceof DLNAMediaInfo) {
 			DLNAMediaInfo mediaCloned = ((DLNAMediaInfo) cloned);
-			mediaCloned.setAudioCodes(new ArrayList<DLNAMediaAudio>());
-			for (DLNAMediaAudio audio : getAudioCodes()) {
-				mediaCloned.getAudioCodes().add((DLNAMediaAudio) audio.clone());
+			mediaCloned.setAudioTracksList(new ArrayList<DLNAMediaAudio>());
+			for (DLNAMediaAudio audio : getAudioTracksList()) {
+				mediaCloned.getAudioTracksList().add((DLNAMediaAudio) audio.clone());
 			}
-			mediaCloned.setSubtitlesCodes(new ArrayList<DLNAMediaSubtitle>());
-			for (DLNAMediaSubtitle sub : getSubtitlesCodes()) {
-				mediaCloned.getSubtitlesCodes().add((DLNAMediaSubtitle) sub.clone());
+			mediaCloned.setSubtitleTracksList(new ArrayList<DLNAMediaSubtitle>());
+			for (DLNAMediaSubtitle sub : getSubtitleTracksList()) {
+				mediaCloned.getSubtitleTracksList().add((DLNAMediaSubtitle) sub.clone());
 			}
 		}
 
@@ -1381,6 +1370,22 @@ public class DLNAMediaInfo implements Cloneable {
 	}
 
 	/**
+	 * @return the frameRateMode
+	 * @since 1.55
+	 */
+	public String getFrameRateMode() {
+		return frameRateMode;
+	}
+
+	/**
+	 * @param frameRateMode the frameRateMode to set
+	 * @since 1.55
+	 */
+	public void setFrameRateMode(String frameRateMode) {
+		this.frameRateMode = frameRateMode;
+	}
+
+	/**
 	 * @return the aspect
 	 * @since 1.50
 	 */
@@ -1445,35 +1450,79 @@ public class DLNAMediaInfo implements Cloneable {
 	}
 
 	/**
-	 * @return the audioCodes
-	 * @since 1.50
+	 * @return the audioTracks
+	 * @since 1.60
 	 */
+	public List<DLNAMediaAudio> getAudioTracksList() {
+		return audioTracks;
+	}
+
+	/**
+	 * @return the audioTracks
+	 * @deprecated use getAudioTracksList() instead
+	 */
+	@Deprecated
 	public ArrayList<DLNAMediaAudio> getAudioCodes() {
-		return audioCodes;
+		if(audioTracks instanceof ArrayList) {
+			return (ArrayList<DLNAMediaAudio>) audioTracks;
+		} else {
+			return new ArrayList<DLNAMediaAudio>();
+		}
 	}
 
 	/**
-	 * @param audioCodes the audioCodes to set
-	 * @since 1.50
+	 * @param audioTracks the audioTracks to set
+	 * @since 1.60
 	 */
-	public void setAudioCodes(ArrayList<DLNAMediaAudio> audioCodes) {
-		this.audioCodes = audioCodes;
+	public void setAudioTracksList(List<DLNAMediaAudio> audioTracks) {
+		this.audioTracks = audioTracks;
 	}
 
 	/**
-	 * @return the subtitlesCodes
-	 * @since 1.50
+	 * @param audioTracks the audioTracks to set
+	 * @deprecated use setAudioTracksList(ArrayList<DLNAMediaAudio> audioTracks) instead
 	 */
+	@Deprecated
+	public void setAudioCodes(List<DLNAMediaAudio> audioTracks) {
+		setAudioTracksList(audioTracks);
+	}
+
+	/**
+	 * @return the subtitleTracks
+	 * @since 1.60
+	 */
+	public List<DLNAMediaSubtitle> getSubtitleTracksList() {
+		return subtitleTracks;
+	}
+
+	/**
+	 * @return the subtitleTracks
+	 * @deprecated use getSubtitleTracksList() instead
+	 */
+	@Deprecated
 	public ArrayList<DLNAMediaSubtitle> getSubtitlesCodes() {
-		return subtitlesCodes;
+		if(subtitleTracks instanceof ArrayList) {
+			return (ArrayList<DLNAMediaSubtitle>) subtitleTracks;
+		} else {
+			return new ArrayList<DLNAMediaSubtitle>();
+		}
 	}
 
 	/**
-	 * @param subtitlesCodes the subtitlesCodes to set
-	 * @since 1.50
+	 * @param subtitleTracks the subtitleTracks to set
+	 * @since 1.60
 	 */
-	public void setSubtitlesCodes(ArrayList<DLNAMediaSubtitle> subtitlesCodes) {
-		this.subtitlesCodes = subtitlesCodes;
+	public void setSubtitleTracksList(List<DLNAMediaSubtitle> subtitleTracks) {
+		this.subtitleTracks = subtitleTracks;
+	}
+
+	/**
+	 * @param subtitleTracks the subtitleTracks to set
+	 * @deprecated use setSubtitleTracksList(ArrayList<DLNAMediaSubtitle> subtitleTracks) instead
+	 */
+	@Deprecated
+	public void setSubtitlesCodes(List<DLNAMediaSubtitle> subtitleTracks) {
+		setSubtitleTracksList(subtitleTracks);
 	}
 
 	/**
